Attempt to make MFA work better
This reimplements some of the flow and improves the behavior when
prompted for an MFA code. This will also make things easier to
extend/fix issues later.

Unfortunately, this code doesn't yet work for the main way we use MFA
with okta, which is per-application rather than when you log into okta.
I'm putting these changes in place anyway because they will make life
easier when we do try to support application based MFA (we will need to
do an action if the result of the okta login is SUCCESS, and you can't
hook into this without your own log_into_okta function).