Remove dependency on oktaauth library
The main change here is to remove the dependency on the external
library for logging into okta and getting the saml assertion. We weren't
able to use it without workarounds, and it prevented us from making
other improvements.

In addition to this, the configuration has changed slightly, instead of
needing an app id, principal_arn and role_arn for every aws account,
only the embed_url is needed. This will be a breaking change, but it
vastly simplifies the configuration and makes it much easier to add new
accounts.

This is achieved by obtaining the saml assertion in the same way the
official okta app does, and then parsing the saml assertion to retrieve
the principal and role ARNs.

One other major change here is the removal of the timer where the
program stays in the foreground. Something will be put back into place
for this, but I want to be able to support the situation where people
don't want the application to take up a terminal window.