Store session cookie and reuse it
This lets you invoke okta_aws multiple times without having to log in to
okta each time, if your okta session from a previous invocation is still
valid. This limits/removes the need for the countdown timer that was
previously in the tool, and it's fairly simple to include a wrapper
script that calls okta_aws once per hour if renewal is needed.

Note: you're still prompted for which role to use on each invocation if
you have more than one role assigned to you. The role configuration
option needs to be respected to make this fully automated.