Change how config file is loaded
Previously, the toml file was loaded, and a single hash containing
configuration merged from the general and profile specific sections was
returned. This assumed that no more than one profile would ever be used
in a single okta_aws run.

This changes the way configuration is loaded to just return the raw
configuration, and provides a get_config helper method for loading
configuration from the right section, based on what the profile is
currently set to.

The current profile is set in self.config, and can be changed during a
single program run, better supporting the --all option.