Restore the AWS_PROFILE environment variable cleaning
Even when using boto3, you run into issues if AWS_PROFILE points to a
profile that doesn't exist. You also run into issues if you pass in a
profile name of 'default' if there is no ~/.aws/config or
~/.aws/credentials file (in other words, if the default profile doesn't
exist). But, if you ensure AWS_PROFILE is blank, and don't pass in a
profile, then boto3 will happily carry on, returning a blank dict when
requesting credential information.